+++ comiccrawler/gui/__init__.py
@@ -1,7 +1,3 @@
 #! python3
 
 """Comic Crawler GUI."""
-
-def main():
-	from .main_window import MainWindow
-	MainWindow()
