+++ requirements.txt
@@ -1,7 +1,7 @@
 docutils==0.15.2
 livereload==2.6.1
 pygments==2.4.2
-pylint==2.3.1
+pylint==2.4.3
 pyxcute==0.5.2
-twine==1.13.0
+twine==2.0.0
 wheel==0.33.6
