+++ setup.cfg
@@ -30,7 +30,7 @@ packages = find:
 
 install_requires =
 	belfrywidgets~=1.0
-	desktop3~=0.5.2
+	desktop3~=0.5.3
 	docopt~=0.6.2
 	node_vm2~=0.3.6
 	pythreadworker~=0.9.0
@@ -39,7 +39,7 @@ install_requires =
 	
 [options.entry_points]
 console_scripts = 
-	comiccrawler = comiccrawler:console_init
+	comiccrawler = comiccrawler.cli:console_init
 
 ; [options.package_data]
 ; node_vm2 =
